requirements.txt
telethon
python-dotenv
requests
PyPDF2
